// resources/js/api/generated/models/serverBackups.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { StackBackupSummary } from './stackBackupSummary';

export interface ServerBackups {
  configured: boolean;
  enabled: boolean;
  error?: string;
  /** @minimum 0 */
  server_id: number;
  server_name: string;
  stacks: StackBackupSummary[];
}