// resources/js/api/generated/models/stackBackupSummary.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { RunSummary } from './runSummary';

export interface StackBackupSummary {
  latest_run?: RunSummary | null;
  /** @minimum 0 */
  repo_size_bytes?: number;
  run_count: number;
  stack_exists: boolean;
  stack_name: string;
}